+++ packages/field/src/components/Select/index.tsx
@@ -30,7 +30,7 @@ import './index.less';
 
 let testId = 0;
 
-export type FieldSelectProps = {
+export type FieldSelectProps<FieldProps = any> = {
   text: string;
   /** 值的枚举，如果存在枚举，Search 中会生成 select */
   valueEnum?: ProFieldValueEnumType;
@@ -41,7 +41,10 @@ export type FieldSelectProps = {
   params?: any;
 
   /** 组件的全局设置 */
-  fieldProps?: any;
+  fieldProps?: FieldProps;
+
+  bordered?: boolean;
+  id?: string;
 };
 
 export const ObjToMap = (value: ProFieldValueEnumType | undefined): ProSchemaValueEnumMap => {
@@ -230,14 +233,14 @@ export const proFieldParsingValueEnumToArray = (
 
     if (typeof value === 'object' && value?.text) {
       enumArray.push({
-        text: (value?.text as unknown) as string,
+        text: value?.text as unknown as string,
         value: key,
         disabled: value.disabled,
       });
       return;
     }
     enumArray.push({
-      text: (value as unknown) as string,
+      text: value as unknown as string,
       value: key,
     });
   });
@@ -378,6 +381,10 @@ const FieldSelect: ProFieldFC<FieldSelectProps> = (props, ref) => {
     children,
     light,
     proFieldKey,
+    params,
+    label,
+    bordered,
+    id,
     ...rest
   } = props;
 
@@ -417,7 +424,7 @@ const FieldSelect: ProFieldFC<FieldSelectProps> = (props, ref) => {
       <>
         {proFieldParsingText(
           rest.text,
-          (ObjToMap(valueEnum || optionsValueEnum) as unknown) as ProSchemaValueEnumObj,
+          ObjToMap(valueEnum || optionsValueEnum) as unknown as ProSchemaValueEnumObj,
         )}
       </>
     );
@@ -433,12 +440,15 @@ const FieldSelect: ProFieldFC<FieldSelectProps> = (props, ref) => {
       if (light) {
         return (
           <LightSelect
+            bordered={bordered}
+            id={id}
             loading={loading}
             ref={inputRef}
             allowClear
             size={size}
-            {...rest}
             options={options}
+            label={label}
+            placeholder={intl.getMessage('tableForm.selectPlaceholder', '请选择')}
             {...fieldProps}
           />
         );
@@ -449,6 +459,8 @@ const FieldSelect: ProFieldFC<FieldSelectProps> = (props, ref) => {
           style={{
             minWidth: 100,
           }}
+          bordered={bordered}
+          id={id}
           loading={loading}
           ref={inputRef}
           allowClear
@@ -464,8 +476,8 @@ const FieldSelect: ProFieldFC<FieldSelectProps> = (props, ref) => {
             }
             return item.label;
           }}
-          {...rest}
           placeholder={intl.getMessage('tableForm.selectPlaceholder', '请选择')}
+          label={label}
           {...fieldProps}
           options={options}
         />