+++ packages/form/src/components/FieldSet/demos/components-other-readonly.tsx
@@ -83,7 +83,16 @@ const Demo = () => (
       }}
       onFinish={async (value) => console.log(value)}
     >
-      <ProFormText readonly width="md" name="name" label="name" />
+      <ProFormText
+        readonly
+        width="md"
+        name="name"
+        label="name"
+        fieldProps={{
+          prefix: 'prefix',
+          suffix: 'suffix',
+        }}
+      />
       <ProFormSelect
         name="select"
         readonly