+++ packages/form/src/demos/components-test.tsx
@@ -10,7 +10,7 @@ import ProForm, {
   ProFormSlider,
   ProFormUploadDragger,
 } from '@ant-design/pro-form';
-import { Input } from 'antd';
+import { Button, Input } from 'antd';
 
 const Demo = () => {
   const formRef = useRef();
@@ -53,6 +53,10 @@ const Demo = () => {
       <ProFormField>
         <Input />
       </ProFormField>
+      <ProForm.Item>
+        <Button>查看记录数</Button>
+        <span>共有200条</span>
+      </ProForm.Item>
     </ProForm>
   );
 };