+++ packages/table/src/components/EditableTable/demos/children.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import type { ProColumns } from '@ant-design/pro-table';
 import { EditableProTable } from '@ant-design/pro-table';
-import ProField from '@ant-design/pro-field';
 import ProCard from '@ant-design/pro-card';
+import { ProFormField } from '@ant-design/pro-form';
 
 const waitTime = (time: number = 100) => {
   return new Promise((resolve) => {
@@ -170,7 +170,8 @@ export default () => {
         }}
       />
       <ProCard title="表格数据" headerBordered collapsible defaultCollapsed>
-        <ProField
+        <ProFormField
+          ignoreFormItem
           fieldProps={{
             style: {
               width: '100%',