+++ packages/descriptions/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.8](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.9.7...@ant-design/pro-descriptions@1.9.8) (2021-08-26)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
+## [1.9.7](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.9.6...@ant-design/pro-descriptions@1.9.7) (2021-08-24)
+
+**Note:** Version bump only for package @ant-design/pro-descriptions
+
 ## [1.9.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-descriptions@1.9.5...@ant-design/pro-descriptions@1.9.6) (2021-08-19)
 
 ### Bug Fixes