+++ packages/list/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.12.2...@ant-design/pro-list@1.12.3) (2021-08-26)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [1.12.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.12.1...@ant-design/pro-list@1.12.2) (2021-08-24)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
 ## [1.12.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.12.0...@ant-design/pro-list@1.12.1) (2021-08-19)
 
 ### Bug Fixes